codegen: compute encoded sizes in u64 so over-limit messages can't wrap
Generated compute_size previously accumulated in u32: three 1.5 GiB
sub-trees wrap to an under-limit value the new entry-point check cannot
distinguish from a legitimate size, and a single huge bytes field wraps in
its usize-to-u32 cast. Accumulate in a u64 local (which cannot overflow
for any in-memory message) and saturate to u32 once at each node's return
via buffa::saturate_size - the trait signature and SizeCache stay u32, and
the byte-exact size survives the 2-4 GiB window so the entry-point guard
sees the true value.

The same discipline applies everywhere size arithmetic lives: write_to's
transient packed-payload and map-entry lengths, the map_codec runtime
helpers (MapCodec::encoded_len, field_len, message_field_len), the
lazy-fragment emitter, and tag lengths. put_len_delimited_header takes a
plain u64 length. Lazy views' inherent encode entry points route through
the shared buffa funnel, gain try_encode / try_encoded_len /
try_encode_to_vec / try_encode_to_bytes twins, and the debug two-pass
byte-count assertion.

A new codegen test locks the discipline by asserting the emitted
compute_size/write_to bodies contain no u32 size arithmetic. Regenerates
the checked-in WKT and bootstrap descriptor types.

+++ buffa/src/types.rs
@@ -638,12 +638,18 @@ put_field_fn!(
 /// length coming from the [`SizeCache`](crate::SizeCache)) and for packed
 /// repeated fields (the payload loop follows).
 ///
-/// The arguments are `(field_number, len)` — both `u32`, so transposing them
-/// compiles but emits a structurally-valid-but-wrong header.
+/// The arguments are `(field_number, len)` — both plain integers, so
+/// transposing them can compile (integer literals infer either width) but
+/// emits a structurally-valid-but-wrong header.
+///
+/// `len` is `u64`, like all size arithmetic feeding the encode path;
+/// callers holding a `u32` (e.g. from
+/// [`SizeCache::consume_next`](crate::SizeCache::consume_next)) widen with
+/// `u64::from`.
 #[inline(always)]
-pub fn put_len_delimited_header(field_number: u32, len: u32, buf: &mut impl BufMut) {
+pub fn put_len_delimited_header(field_number: u32, len: u64, buf: &mut impl BufMut) {
     Tag::new(field_number, WireType::LengthDelimited).encode(buf);
-    encode_varint(len as u64, buf);
+    encode_varint(len, buf);
 }
 
 /// Write a group field's `StartGroup` tag.
